agents/pydantic-weather-agent/requirements.txt
pydantic_ai
pydantic-ai-slim[duckduckgo]
python-dotenv
tavily-python